Fix optional pausing when focus is lost
This was broken for Emscripten builds because of use of
SDL_WaitEvent(). It also prevented SDL 1 builds because that
function is missing.

This unifies SDL 1 and SDL 2 code for pausing when focus is lost,
and creates an Emscripten version similar to ordinary pause.
Emscripten SDL 1 does not seem to deliver these events, so use
SDL 2 if you want to use this feature. It is disabled by default
and can be enabled via the configuration file:
[sdl]
priority=higher,pause