Fix resume from loss of focus pause, avoiding slowdown
When paused due to loss of focus, events up to and including
regaining of focus are delivered to another place. The code there
didn't do some of the things that the main event code does after
regaining focus. The most important thing that was missed was
re-enabling CPU cycle auto adjustment. This bug existed in
original code, and was preserved as it was restructured for
Emscripten.